Propagation code is now working.
We should now be OK with both variable diameters and variable lengths.
Only variable lengths have been subjected to anything resembling a
test. The expression sigma is actually coming directly from Aldo's PhD
thesis, and is a generalisation of the expression used in this file
before (2 commits ago).
It is straightforward to support variable Ra, and I should remove the
tests on constant diameter and constant Ra.